Move inRange.js to assessment helpers
