+++ packages/yoastseo/src/scoring/assessments/readability/passiveVoiceAssessment.js
@@ -2,7 +2,7 @@ import { map } from "lodash-es";
 
 import formatNumber from "../../../helpers/formatNumber";
 import getLanguageAvailability from "../../../helpers/_todo/getLanguageAvailability";
-import { inRangeEndInclusive as inRange } from "../../../helpers/inRange";
+import { inRangeEndInclusive as inRange } from "../../helpers/assessments/inRange";
 import marker from "../../../markers/addMark";
 import { createAnchorOpeningTag } from "../../../helpers/shortlinker";
 import { stripIncompleteTags as stripTags } from "../../../../researches/stringProcessing/stripHTMLTags";
