+++ packages/yoastseo/src/scoring/assessments/readability/sentenceLengthInDescriptionAssessment.js
@@ -1,5 +1,5 @@
 import formatNumber from "../../../helpers/formatNumber";
-import { inRangeEndInclusive as inRange } from "../../../helpers/inRange";
+import { inRangeEndInclusive as inRange } from "../../helpers/assessments/inRange";
 import { createAnchorOpeningTag } from "../../../helpers/shortlinker";
 import AssessmentResult from "../../../values/AssessmentResult";
 import countTooLongSentences from "../../helpers/assessments/checkForTooLongSentences";
