+++ packages/yoastseo/src/scoring/assessments/seo/TextImagesAssessment.js
@@ -1,7 +1,7 @@
 import { merge } from "lodash-es";
 
 import Assessment from "../assessment";
-import { inRangeStartEndInclusive } from "../../../helpers/inRange.js";
+import { inRangeStartEndInclusive } from "../../helpers/assessments/inRange.js";
 import { createAnchorOpeningTag } from "../../../helpers/shortlinker";
 import AssessmentResult from "../../../values/AssessmentResult";
